[fix] utilに切り分ける
